+++ R/pkgbuild.R
@@ -3,9 +3,7 @@
 #' @export
 build <- function(path = ".", dest_path = NULL, binary = FALSE, vignettes = TRUE,
                   manual = FALSE, args = NULL, quiet = FALSE, ...) {
-  if (rstudioapi::hasFun("documentSaveAll")) {
-    rstudioapi::documentSaveAll()
-  }
+  save_all()
   pkgbuild::build(path = path, dest_path = dest_path, binary = binary,
     vignettes = vignettes, manual = manual, args = args, quiet = quiet, ...)
 }